fix_makefile & errors

+++ cpp00/ex00/megaphone.cpp
@@ -2,7 +2,7 @@
 
 std::string toUpperStr(std::string s)
 {
-    for (int i = 0; i < (int)s.length(); i++)
+    for (size_t i = 0; i < s.length(); i++)
         s[i] = toupper(s[i]);
     return s;
 }
